requirements.txt
flask==2.3.2
torch==2.0.1
torchvision==0.15.2
transformers==4.30.0
pillow==9.5.0
onnxruntime>=1.15.0
onnx>=1.17.0
